More STFreader changes.
Now handles the situation where the last item in a block is an ignored item (skip comment #* or _*), which was tripping up the more simplistic test in EndOfBlock().
If EndOfBlock() returns false it should guarantee that your next call to ReadItem() will not return a ")" - now that "contract" is always honoured

git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@425 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/MSTS/CVFFile.cs
@@ -236,8 +236,8 @@ namespace MSTS
             else if (string.Compare(token, "ScaleRange", true) == 0)
             {
                 inf.MustMatch("(");
-                MinValue = inf.ReadInt(STFReader.UNITS.Any, null);
-                MaxValue = inf.ReadInt(STFReader.UNITS.Any, null);
+                MinValue = inf.ReadDouble(STFReader.UNITS.Any, null);
+                MaxValue = inf.ReadDouble(STFReader.UNITS.Any, null);
                 inf.SkipRestOfBlock();
             }
             else if (string.Compare(token, "Graphic", true) == 0)
